+++ tests/test_settings.py
@@ -18,12 +18,6 @@ def test_load_settings_file_absent(settings_default, settings_file_absent):
     assert settings.dict() == settings_default
 
 
-def test_load_settings_posix_tmp(settings_posix_tmp, settings_file_posix_tmp):
-    """Test loading settings with working settings file."""
-    settings = load_settings(settings_file_posix_tmp)
-    assert settings.dict() == settings_posix_tmp
-
-
 def test_load_settings_file_malformed(settings_file_malformed):
     """Test raising exception when settings file is corrupted."""
     with pytest.raises(FaddrSettingsFileFormatError):